Re-organize components into proper groups.
