+++ resources/assets/Application.js
@@ -2,14 +2,14 @@ import React from 'react';
 import { BrowserRouter, Route, Switch, Redirect } from 'react-router-dom';
 import { ApolloProvider } from '@apollo/react-hooks';
 
-import { env } from '../helpers';
-import graphql from '../graphql';
-import ShowPost from './ShowPost';
-import UserIndex from './UserIndex';
-import ShowSignup from './ShowSignup';
-import ShowCampaign from './ShowCampaign';
-import UserOverview from './UserOverview';
-import CampaignIndex from './CampaignIndex';
+import { env } from './helpers';
+import graphql from './graphql';
+import ShowPost from './pages/ShowPost';
+import ShowUser from './pages/ShowUser';
+import UserIndex from './pages/UserIndex';
+import ShowSignup from './pages/ShowSignup';
+import ShowCampaign from './pages/ShowCampaign';
+import CampaignIndex from './pages/CampaignIndex';
 
 const Application = () => {
   const endpoint = env('GRAPHQL_URL');
@@ -30,7 +30,7 @@ const Application = () => {
             <UserIndex />
           </Route>
           <Route path="/users/:id">
-            <UserOverview />
+            <ShowUser />
           </Route>
           <Route path="/posts/:id">
             <ShowPost />
