+++ resources/assets/components/DeleteSignupButton.js
