+++ resources/assets/components/utilities/MetaInformation.js
