+++ resources/assets/components/utilities/Quantity.js
