+++ resources/assets/pages/UserIndex.js
@@ -3,8 +3,8 @@ import { Link } from 'react-router-dom';
 import { useQuery } from '@apollo/react-hooks';
 import React, { useRef, useState } from 'react';
 
-import Empty from './Empty';
-import Shell from './utilities/Shell';
+import Empty from '../components/Empty';
+import Shell from '../components/utilities/Shell';
 
 const USER_SEARCH_QUERY = gql`
   query UserSearchQuery($term: String!) {
